bing-search-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.5
httpx==0.27.0
pydantic==2.9.0
python-dotenv==1.0.1
typer==0.12.4
